+++ src/min/stoch/ag.hpp
@@ -54,7 +54,7 @@ namespace min
                 ///
                 tstate operator()(const tproblem& problem, const tvector& x0) const
                 {
-                        assert(problem.size() == static_cast<std::size_t>(x0.size()));
+                        assert(problem.size() == x0.size());
 
                         // restart method
                         trestart restart;
